project/requirements.txt
opencv-python==4.8.0.74
numpy==1.26.4
wheel
tqdm
pytorch_lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy
imageio
lpips
torchtyping
pillow
wandb-osh
lightning